Conflicts:
	com/as.infrastructure/system/kernel/contiki/core/os_i.h
	com/as.infrastructure/system/kernel/contiki/platform/native/portable.c
	com/as.tool/config.infrastructure.system/argen/gen_contiki.py
	release/make/posix.mk

+++ com/as.infrastructure/communication/RPmsg/virtio_ring.h
@@ -55,7 +55,7 @@
 /* Virtio ring descriptors: 16 bytes.  These can chain together via "next". */
 typedef struct {
 	/* Address (guest-physical). */
-	uint32 addr;
+	uint64 addr;
 	/* Length. */
 	uint32 len;
 	/* The flags as indicated above. */
